move LinearRelationship into module
